pass more of mypy strict mode.
the main thing keeping us from 100% strict mode is the lack of typing for the interfaces of the libraries we use (I haven't checked all possible avenues of acquiring these types; I don't know what a typeshed is yet.) but that just needs a flag to ignore. This change mostly consisted in adding additonal type annotations, but also the unused helper function create_tf_serving_json was removed.

Also, bump databricks-sql-connector version to 3.0.1, out of a mistaken belief that there might be types in the interface of this one. Ah, well.