remove dataframe step from the send process
which turned out to be redundant, as the endpoint just converts it to a dataframe or something internally anyway. However, I was not able to unpsycho the dict we send because apparently it's a loadbearing assumption of mlflow that you're sending a table of prompts, not just a single prompt. So we have to just make everything a list of one, whatever