try, apparently unsuccessfully, to fix the thing where Yeterday doesn't work
oddly, the other days work, and also filtering for yesterday seems to work in the databricks query. I thought maybe it was a fencepost error, not playing well with the new actually-a-date date data type, and that helped a little, but it still seems to be wrong. I also thought maybe the accounts were off, but couldn't fix it that way, either.