Use pages instead of tabs.
(Perf improvement, following Streamlit improvement to pages.)

Every vertical pixel used for the UI bar signifies a theft from those who hunger and are not fed, those who are cold and are not clothed.

—https://x.com/admittedlyhuman/status/430532517389561856

Here is the message I wrote earlier about this plan, before I knew it would work:

Streamlit recently made some changes to how pages work, which I'm hoping means they will no longer drop all the UI state on the floor every page switch. So I will evaluate actual pages instead of tabs.

This would allow us to remove code, improve performance slightly, and also allow us to remove some code that we currently have to improve performance in the tab paradigm, so it could be a sizable but ultimately minor win.

It might ALSO allow us to use the streaming output version of the chatbot, which might improve the user experience of performance (since I think the streaming does literally get the first tokens in faster, so the users can start reading faster. If the streaming were just a graphical effect I would say it was nonsense, but my uninformed impression is that they really are taking a while to generate each token and sending us the tokens as they come.)