handle the index error
OK well... the answer to the question of why we're getting that error all the time is that someone is logged in as 'test@test.com'. Somewhere. This forces my hand into actually handling the circumstance, even though my instinct is instead to try to make the incorrection circumstance impossible