CHAOTIC NEUTRAL COMMIT: turn off health check
I guess the existence of this option means we could turn off the health check, potentially fixing that problem we were having (and no longer have currently) where having the app around was keeping the model server alive and costing us thousands of dollars. The consequences are difficult to predict. Like, maybe it will no longer automatically reload if the app ever goes down. But let's do it lmao. (All agreed to this.)